487: Refactoring Internation Payments tests
Applying the test refactoring to allow logic to be reused to call different API versions

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/487
