+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/payments/consents/legacy/LegacyCreateInternationalPaymentsConsentsTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.consents
+package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.consents.legacy
 
 import assertk.assertThat
 import assertk.assertions.contains
@@ -24,206 +24,13 @@ import com.github.kittinunf.fuel.core.FuelError
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
-import uk.org.openbanking.datamodel.payment.*
+import uk.org.openbanking.datamodel.payment.OBWriteInternationalConsentResponse2
+import uk.org.openbanking.datamodel.payment.OBWriteInternationalConsentResponse3
+import uk.org.openbanking.datamodel.payment.OBWriteInternationalConsentResponse4
+import uk.org.openbanking.datamodel.payment.OBWriteInternationalConsentResponse5
 import uk.org.openbanking.testsupport.payment.OBWriteInternationalConsentTestDataFactory.*
 
-class CreateInternationalPaymentsConsentsTest(val tppResource: CreateTppCallback.TppResource) {
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateInternationalPaymentConsent"],
-        apis = ["international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun createInternationalPaymentsConsents_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        // When
-        val consent = PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-            payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        // Then
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-        assertThat(consent.risk).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateInternationalPaymentConsent"],
-        apis = ["international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun createInternationalPaymentsConsents_mandatoryFields_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5MandatoryFields()
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        // When
-        val consent = PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-            payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        // Then
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-        assertThat(consent.risk).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateInternationalPaymentConsent"],
-        apis = ["international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateInternationalPaymentConsents_throwsSendInvalidFormatDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-                payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8,
-                INVALID_FORMAT_DETACHED_JWS
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
-        assertThat(exception.message.toString()).contains(INVALID_FORMAT_DETACHED_JWS_ERROR)
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateInternationalPaymentConsent"],
-        apis = ["international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateInternationalPaymentConsents_throwsNoDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consentNoDetachedJwt<OBWriteInternationalConsentResponse6>(
-                payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
-        assertThat(exception.message.toString()).contains(NO_DETACHED_JWS)
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateInternationalPaymentConsent"],
-        apis = ["international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateInternationalPaymentConsents_throwsNotPermittedB64HeaderAddedInTheDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-
-        val signedPayload =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid,
-                true
-            )
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-                payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8,
-                signedPayload
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.responseMessage).isEqualTo(UNAUTHORIZED)
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(401)
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateInternationalPaymentConsent"],
-        apis = ["international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateInternationalPaymentConsents_throwsSendInvalidKidDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                INVALID_SIGNING_KID
-            )
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-                payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8,
-                signedPayloadConsent
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(401)
-        assertThat((exception.cause as FuelError).response.responseMessage).isEqualTo(UNAUTHORIZED)
-    }
-
+class LegacyCreateInternationalPaymentsConsentsTest(val tppResource: CreateTppCallback.TppResource) {
     @EnabledIfVersion(
         type = "payments",
         apiVersion = "v3.1.4",
