+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/payments/legacy/LegacyGetInternationalPaymentInternationalPaymentIdPaymentDetailsTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.international.payments
+package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.legacy
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
@@ -22,412 +22,7 @@ import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.payment.*
 import uk.org.openbanking.testsupport.payment.OBWriteInternationalConsentTestDataFactory.*
 
-class GetInternationalPaymentInternationalPaymentIdPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_AGREED_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-        consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.AGREED
-
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        val consent = PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-            payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        assertThat(consent.data.initiation.exchangeRateInformation.exchangeRate).isNotNull()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-
-        // accessToken to submit payment use the grant type authorization_code
-        val accessTokenAuthorizationCode = PaymentAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        // accessToken to get the payment use the grant type client_credentials
-        val accessTokenClientCredentials = PaymentRS().getAccessToken(tppResource.tpp)
-
-        val patchedConsent = PaymentRS().getConsent<OBWriteInternationalConsentResponse6>(
-            PaymentFactory.urlWithConsentId(
-                payment3_1_8.Links.links.GetInternationalPaymentConsent,
-                consent.data.consentId
-            ),
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(patchedConsent).isNotNull()
-        assertThat(patchedConsent.data).isNotNull()
-        assertThat(patchedConsent.risk).isNotNull()
-        assertThat(patchedConsent.data.consentId).isNotEmpty()
-        Assertions.assertThat(patchedConsent.data.status.toString()).`is`(Status.consentCondition)
-
-        val standingOrderSubmissionRequest = OBWriteInternational3().data(
-            OBWriteInternational3Data()
-                .consentId(patchedConsent.data.consentId)
-                .initiation(patchedConsent.data.initiation)
-        ).risk(patchedConsent.risk)
-
-        val signedPayload = signPayloadSubmitPayment(
-            defaultMapper.writeValueAsString(standingOrderSubmissionRequest),
-            tppResource.tpp.signingKey,
-            tppResource.tpp.signingKid
-        )
-
-        val submissionResponse = PaymentRS().submitPayment<OBWriteInternationalResponse5>(
-            payment3_1_8.Links.links.CreateInternationalPayment,
-            standingOrderSubmissionRequest,
-            accessTokenAuthorizationCode,
-            signedPayload,
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(submissionResponse).isNotNull()
-        assertThat(submissionResponse.data).isNotNull()
-        assertThat(submissionResponse.data.consentId).isEqualTo(patchedConsent.data.consentId)
-        assertThat(submissionResponse.data.internationalPaymentId).isNotEmpty()
-
-        // When
-        val result = PaymentRS().getPayment<OBWritePaymentDetailsResponse1>(
-            urlWithInternationalPaymentId(
-                payment3_1_8.Links.links.GetInternationalPaymentInternationalPaymentIdPaymentDetails,
-                submissionResponse.data.internationalPaymentId
-            ),
-            accessTokenClientCredentials,
-            tppResource.tpp
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data).isNotNull()
-        assertThat(result.data.paymentStatus).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_ACTUAL_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-        consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.ACTUAL
-        consentRequest.data.initiation.exchangeRateInformation.exchangeRate = null
-        consentRequest.data.initiation.exchangeRateInformation.contractIdentification = null
-
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        val consent = PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-            payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-
-        // accessToken to submit payment use the grant type authorization_code
-        val accessTokenAuthorizationCode = PaymentAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        // accessToken to get the payment use the grant type client_credentials
-        val accessTokenClientCredentials = PaymentRS().getAccessToken(tppResource.tpp)
-
-        val patchedConsent = PaymentRS().getConsent<OBWriteInternationalConsentResponse6>(
-            PaymentFactory.urlWithConsentId(
-                payment3_1_8.Links.links.GetInternationalPaymentConsent,
-                consent.data.consentId
-            ),
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(patchedConsent).isNotNull()
-        assertThat(patchedConsent.data).isNotNull()
-        assertThat(patchedConsent.risk).isNotNull()
-        assertThat(patchedConsent.data.consentId).isNotEmpty()
-        Assertions.assertThat(patchedConsent.data.status.toString()).`is`(Status.consentCondition)
-
-        val standingOrderSubmissionRequest = OBWriteInternational3().data(
-            OBWriteInternational3Data()
-                .consentId(patchedConsent.data.consentId)
-                .initiation(patchedConsent.data.initiation)
-        ).risk(patchedConsent.risk)
-
-        val signedPayload = signPayloadSubmitPayment(
-            defaultMapper.writeValueAsString(standingOrderSubmissionRequest),
-            tppResource.tpp.signingKey,
-            tppResource.tpp.signingKid
-        )
-
-        val submissionResponse = PaymentRS().submitPayment<OBWriteInternationalResponse5>(
-            payment3_1_8.Links.links.CreateInternationalPayment,
-            standingOrderSubmissionRequest,
-            accessTokenAuthorizationCode,
-            signedPayload,
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(submissionResponse).isNotNull()
-        assertThat(submissionResponse.data).isNotNull()
-        assertThat(submissionResponse.data.consentId).isEqualTo(patchedConsent.data.consentId)
-        assertThat(submissionResponse.data.internationalPaymentId).isNotEmpty()
-
-        // When
-        val result = PaymentRS().getPayment<OBWritePaymentDetailsResponse1>(
-            urlWithInternationalPaymentId(
-                payment3_1_8.Links.links.GetInternationalPaymentInternationalPaymentIdPaymentDetails,
-                submissionResponse.data.internationalPaymentId
-            ),
-            accessTokenClientCredentials,
-            tppResource.tpp
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data).isNotNull()
-        assertThat(result.data.paymentStatus).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_INDICATIVE_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5()
-        consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.INDICATIVE
-        consentRequest.data.initiation.exchangeRateInformation.exchangeRate = null
-        consentRequest.data.initiation.exchangeRateInformation.contractIdentification = null
-
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        val consent = PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-            payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-
-        // accessToken to submit payment use the grant type authorization_code
-        val accessTokenAuthorizationCode = PaymentAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        // accessToken to get the payment use the grant type client_credentials
-        val accessTokenClientCredentials = PaymentRS().getAccessToken(tppResource.tpp)
-
-        val patchedConsent = PaymentRS().getConsent<OBWriteInternationalConsentResponse6>(
-            PaymentFactory.urlWithConsentId(
-                payment3_1_8.Links.links.GetInternationalPaymentConsent,
-                consent.data.consentId
-            ),
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(patchedConsent).isNotNull()
-        assertThat(patchedConsent.data).isNotNull()
-        assertThat(patchedConsent.risk).isNotNull()
-        assertThat(patchedConsent.data.consentId).isNotEmpty()
-        Assertions.assertThat(patchedConsent.data.status.toString()).`is`(Status.consentCondition)
-
-        val standingOrderSubmissionRequest = OBWriteInternational3().data(
-            OBWriteInternational3Data()
-                .consentId(patchedConsent.data.consentId)
-                .initiation(patchedConsent.data.initiation)
-        ).risk(patchedConsent.risk)
-
-        val signedPayload = signPayloadSubmitPayment(
-            defaultMapper.writeValueAsString(standingOrderSubmissionRequest),
-            tppResource.tpp.signingKey,
-            tppResource.tpp.signingKid
-        )
-
-        val submissionResponse = PaymentRS().submitPayment<OBWriteInternationalResponse5>(
-            payment3_1_8.Links.links.CreateInternationalPayment,
-            standingOrderSubmissionRequest,
-            accessTokenAuthorizationCode,
-            signedPayload,
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(submissionResponse).isNotNull()
-        assertThat(submissionResponse.data).isNotNull()
-        assertThat(submissionResponse.data.consentId).isEqualTo(patchedConsent.data.consentId)
-        assertThat(submissionResponse.data.internationalPaymentId).isNotEmpty()
-
-        // When
-        val result = PaymentRS().getPayment<OBWritePaymentDetailsResponse1>(
-            urlWithInternationalPaymentId(
-                payment3_1_8.Links.links.GetInternationalPaymentInternationalPaymentIdPaymentDetails,
-                submissionResponse.data.internationalPaymentId
-            ),
-            accessTokenClientCredentials,
-            tppResource.tpp
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data).isNotNull()
-        assertThat(result.data.paymentStatus).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_v3_1_8_mandatoryFields() {
-        // Given
-        val consentRequest = aValidOBWriteInternationalConsent5MandatoryFields()
-
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        val consent = PaymentRS().consent<OBWriteInternationalConsentResponse6>(
-            payment3_1_8.Links.links.CreateInternationalPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-
-        // accessToken to submit payment use the grant type authorization_code
-        val accessTokenAuthorizationCode = PaymentAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        // accessToken to get the payment use the grant type client_credentials
-        val accessTokenClientCredentials = PaymentRS().getAccessToken(tppResource.tpp)
-
-        val patchedConsent = PaymentRS().getConsent<OBWriteInternationalConsentResponse6>(
-            PaymentFactory.urlWithConsentId(
-                payment3_1_8.Links.links.GetInternationalPaymentConsent,
-                consent.data.consentId
-            ),
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(patchedConsent).isNotNull()
-        assertThat(patchedConsent.data).isNotNull()
-        assertThat(patchedConsent.risk).isNotNull()
-        assertThat(patchedConsent.data.consentId).isNotEmpty()
-        Assertions.assertThat(patchedConsent.data.status.toString()).`is`(Status.consentCondition)
-
-        val standingOrderSubmissionRequest = OBWriteInternational3().data(
-            OBWriteInternational3Data()
-                .consentId(patchedConsent.data.consentId)
-                .initiation(patchedConsent.data.initiation)
-        ).risk(patchedConsent.risk)
-
-        val signedPayload = signPayloadSubmitPayment(
-            defaultMapper.writeValueAsString(standingOrderSubmissionRequest),
-            tppResource.tpp.signingKey,
-            tppResource.tpp.signingKid
-        )
-
-        val submissionResponse = PaymentRS().submitPayment<OBWriteInternationalResponse5>(
-            payment3_1_8.Links.links.CreateInternationalPayment,
-            standingOrderSubmissionRequest,
-            accessTokenAuthorizationCode,
-            signedPayload,
-            tppResource.tpp,
-            OBVersion.v3_1_8
-        )
-
-        assertThat(submissionResponse).isNotNull()
-        assertThat(submissionResponse.data).isNotNull()
-        assertThat(submissionResponse.data.consentId).isEqualTo(patchedConsent.data.consentId)
-        assertThat(submissionResponse.data.internationalPaymentId).isNotEmpty()
-
-        // When
-        val result = PaymentRS().getPayment<OBWritePaymentDetailsResponse1>(
-            urlWithInternationalPaymentId(
-                payment3_1_8.Links.links.GetInternationalPaymentInternationalPaymentIdPaymentDetails,
-                submissionResponse.data.internationalPaymentId
-            ),
-            accessTokenClientCredentials,
-            tppResource.tpp
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data).isNotNull()
-        assertThat(result.data.paymentStatus).isNotNull()
-    }
-
+class LegacyGetInternationalPaymentInternationalPaymentIdPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
     @EnabledIfVersion(
         type = "payments",
         apiVersion = "v3.1.4",
